+++ src/stories/Page.stories.js
@@ -2,10 +2,9 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import StoryRouter from 'storybook-react-router';
 import { withKnobs, object } from '@storybook/addon-knobs/react';
-import ThemeProvider from '@material-ui/styles/ThemeProvider';
-import { linnTheme } from '../themes/linnTheme';
 import Page from '../components/Page';
 import Title from '../components/Title';
+import providers from './renderUtils/Providers';
 
 const props = {
     history: {
@@ -18,12 +17,7 @@ const props = {
 
 const stories = storiesOf('Page', module);
 stories.addDecorator(withKnobs);
-stories.addDecorator(StoryRouter());
-stories.addDecorator(story => (
-    <ThemeProvider theme={linnTheme}>
-        <div>{story()}</div>
-    </ThemeProvider>
-));
+stories.addDecorator(StoryRouter()).addDecorator(story => providers(story));
 
 stories.add('default', () => (
     <Page history={object('location', props.history)}>