+++ CHANGELOG.md
@@ -5,6 +5,16 @@
 
 -   ProcessActions, ProcessSelectors and processStoreFactory
 
+## [7.2.0] - 2019-10-08
+
+### Added
+
+-   New Component - ValidatedInputDialog
+
+### Changed
+
+- Fixed broken stories.
+
 ## [7.1.0] - 2019-10-04
 
 ### Added